Add files via upload
Sleigh Rush is a game that was created for Mitsubishi Electric's Christmas contest. Written in C# and using the .NET Framework. The purpose of this game is to control Santa and avoid objects that are coming towards him using directional keys. This game was uploaded into a Computer Numeric Control simulator machine that had a Windows Operating System. Finished in 1st in the contest.