Add files via upload
This is the Agilant App I developed for my Logistics Coordinate at the time. It will read in bar code scans with a bar code scanner. The values scanned from the asset tags will be stored in a table and a database that was created in MYSQL workbench. AWS was the service that hosted the Relational Database. However, the RDM has been removed and the program will not work.