Add files via upload
This project is a website that was created for the Software Engineer course. It is an auto-part website that users can select and purchase. It was written in HTML , PHP, and C++. In the readme will provide a link of the site. The site should still be up and running.